Access SplashProgress font renderer via mixins
