+++ src/main/java/com/gtnewhorizons/angelica/client/font/SplashFontRendererRef.java
@@ -1,0 +1,9 @@
+package com.gtnewhorizons.angelica.client.font;
+
+import net.minecraft.client.gui.FontRenderer;
+
+public class SplashFontRendererRef {
+    
+    public static FontRenderer fontRenderer;
+
+}
