+++ src/main/java/com/gtnewhorizons/angelica/mixins/Mixins.java
@@ -87,6 +87,7 @@ public enum Mixins {
     ARCHAIC_SPLASH(new Builder("ArchaicFix Splash").addTargetedMod(TargetedMod.VANILLA).setSide(Side.CLIENT)
         .setPhase(Phase.EARLY).setApplyIf(() -> AngelicaConfig.showSplashMemoryBar && !AngelicaMod.lwjglDebug).addMixinClasses(
               "angelica.archaic.MixinSplashProgress"
+             ,"angelica.archaic.MixinSplashProgress$3"
              ,"angelica.archaic.AccessorSplashProgress"
         )
     ),
